+++ framework/internal/src/index.ts
@@ -1,5 +1,5 @@
 export * from './identityMappers';
 export * from './schemaEquality';
 export * from './serviceSchemaResolver';
-export * from './transformIntoInternalDtoMapper';
+export * from './transformIntoInternalMapper';
 export * from './types';