+++ framework/internal/src/types/index.ts
@@ -1,3 +1,3 @@
-export * from './internalDtoMapper.types';
+export * from './internalMapper.types';
 export * from './mappers.types';
 export * from './serviceSchemaResolver.types';