+++ framework/hyper-express/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @forklaunch/hyper-express
 
+## 0.5.32
+
+### Patch Changes
+
+- change dtoMapper to Mapper
+- Updated dependencies
+  - @forklaunch/validator@0.6.16
+  - @forklaunch/common@0.3.14
+  - @forklaunch/core@0.9.21
+
 ## 0.5.31
 
 ### Patch Changes